requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.3
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
docx2txt==0.9
emoji==2.15.0
faiss-cpu==1.8.0
filetype==1.2.0
Flask==3.0.0
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
greenlet==3.3.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.43
langchain-openai==0.1.16
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.147
limits==5.8.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
nltk==3.9.2
numpy==1.26.4
openai==1.109.1
ordered-set==4.1.0
orjson==3.11.7
packaging==24.2
pandas==2.2.2
pgvector==0.2.5
propcache==0.4.1
psycopg2-binary==2.9.9
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
python-iso639==2026.1.31
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.66.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.14.9
unstructured-client==0.42.3
urllib3==2.6.3
Werkzeug==3.0.1
wrapt==2.1.1
yarl==1.22.0
